:earth_africa: BIG update: Add a gobject handler in utilsCSFML lib (not finished), Move utils2 in basic_c lib, Adjusting some basic_c functions with size_t type

+++ lib/basic_c/source/my_putstr.c
@@ -14,7 +14,7 @@
 ** @param	str	String to print on the stdout
 ** @return	The length of {str}
 */
-int my_putstr(char const *str)
+size_t my_putstr(char const *str)
 {
 	int str_len = my_strlen(str);
